修复Bug / 修复 Web 权重检索器 / 稳定测

+++ src/llm_kira/client/types.py
@@ -58,7 +58,7 @@ class Interaction(BaseModel):
 
 class InteractionWeight(BaseModel):
     interaction: Interaction
-    weight: List[int] = []
+    weight: List[float] = []
 
     @property
     def score(self):
